glove: move button_t class to a separate file, add documentation comments
